various adjusts to the system

+++ front/src/pages/Produtos/Produtos.tsx
@@ -6,10 +6,18 @@ import Papa from 'papaparse';
 const Produtos: React.FC = () => {
     const [produtos, setProdutos] = useState<Product[]>([]);
     const [filtroName, setFiltroName] = useState('');
+    const [ocultarZero, setOcultarZero] = useState(false);
+    const [exibeZero, setExibeZero] = useState(false);
+    const [valorTotal, setValorTotal] = useState(0);
+    const [filtroCode, setFiltroCode] = useState('');
 
     useEffect(() => {
         fetchProducts();
     }, []);
+
+    useEffect(() => {
+        calculaValorTotal();
+    }, [produtos]);
 
     const fetchProducts = async () => {
         try {
@@ -24,7 +32,33 @@ const Produtos: React.FC = () => {
         setFiltroName(event.target.value);
     };
 
+    const handleFilterCodeChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+        setFiltroCode(event.target.value);
+    }
+
+    const handleToggleOcultarZero = () => {
+        setOcultarZero(!ocultarZero);
+        setExibeZero(false);
+    }
+
+    const handleToggleExibeZero = () => {
+        setExibeZero(!exibeZero);
+        setOcultarZero(false);
+    }
+
+    const produtosFiltrados = produtos
+        .filter((product: Product) =>
+            product.description.toUpperCase().includes(filtroName.toUpperCase()) &&
+            product.code.toString().includes(filtroCode)
+        )
+        .filter(product => (ocultarZero ? product.quantity > 0 : true))
+        .filter(product => (exibeZero ? product.quantity === 0 : true))
+        .sort((a: Product, b: Product) =>
+            a.description.localeCompare(b.description)
+        )
+
     const handleDeleteProduct = async (id: number) => {
+        if (!window.confirm('Deseja realmente excluir este produto?')) return;
         try {
             await api.delete(`/products/${id}`);
             fetchProducts();
@@ -38,6 +72,7 @@ const Produtos: React.FC = () => {
     const [unit_price, setUnit_price] = useState(0.0);
     const [quantity, setQuantity] = useState(0);
     const [obs, setObs] = useState('');
+    const [ean, setEAN] = useState('');
 
     const handleCreateProduct = async (e: FormEvent) => {
         e.preventDefault();
@@ -47,11 +82,19 @@ const Produtos: React.FC = () => {
                 description,
                 unit_price,
                 quantity,
-                obs
+                obs,
+                ean
             };
             await api.post('/products/', data);
             console.log(data);
             fetchProducts();
+
+            //limpa os campos do produto
+            setCode('');
+            setDescription('');
+            setUnit_price(0.0);
+            setQuantity(0);
+            setObs('');
         } catch (error) {
             console.error(error);
         }
@@ -65,6 +108,7 @@ const Produtos: React.FC = () => {
         setUnit_price(product.unit_price);
         setQuantity(product.quantity);
         setObs(product.obs);
+        setEAN(product.ean);
     };
 
     const clearProduct = () => {
@@ -74,17 +118,23 @@ const Produtos: React.FC = () => {
         setUnit_price(0.0);
         setQuantity(0);
         setObs('');
+        setEAN('');
     }
 
     const handleUpdateProduct = async () => {
         if (!currentProduct) return;
+        if (quantity == 0) {
+            alert('Quantidade não pode ser 0');
+            return;
+        }
         try {
             const updatedProduct = {
                 code,
                 description,
                 unit_price,
                 quantity,
-                obs
+                obs,
+                ean
             };
             await api.put(`/products/${currentProduct.id}`, updatedProduct);
             fetchProducts();
@@ -95,28 +145,38 @@ const Produtos: React.FC = () => {
     };
 
     const handleExportProducts = async () => {
-        const csvData = produtos.map(({ code, description, unit_price, quantity, obs }) => ({
+        const csvData = produtos.map(({ ean, description, unit_price, quantity, obs, code }) => ({
+            EAN: ean,
             CODIGO: code,
             DESCRICAO: description,
             QUANTIDADE: quantity.toString().replace('.', ','),
             VALOR_UNITARIO: unit_price.toString().replace('.', ','),
-            VALOR_TOTAL: (quantity * unit_price).toFixed(2).replace('.', ','),
             OBSERVACAO: obs
         }));
-        
+
         const csv = Papa.unparse(csvData, { delimiter: ';' });
 
         const csvBlob = new Blob([`\uFEFF${csv}`], { type: 'text/csv;charset=utf-8;' });
 
-		const csvURL = window.URL.createObjectURL(csvBlob);
-		const tempLink = document.createElement('a');
-		tempLink.href = csvURL;
-		tempLink.setAttribute('download', 'relatorio.csv');
-		document.body.appendChild(tempLink);
-		tempLink.click();
-		document.body.removeChild(tempLink);
+        const csvURL = window.URL.createObjectURL(csvBlob);
+        const tempLink = document.createElement('a');
+        tempLink.href = csvURL;
+        tempLink.setAttribute('download', 'relatorio.csv');
+        document.body.appendChild(tempLink);
+        tempLink.click();
+        document.body.removeChild(tempLink);
     }
-            
+
+    const formattedPrice = (price: number) => {
+        return new Intl.NumberFormat('pt-BR', { style: 'currency', currency: 'BRL' }).format(price);
+    }
+
+    const calculaValorTotal = () => {
+        let total = 0;
+        produtosFiltrados.map((product) => total += product.unit_price * product.quantity);
+        setValorTotal(total);
+    }
+
 
     return (
         <div className="container">
@@ -129,6 +189,10 @@ const Produtos: React.FC = () => {
                 <div className="col-sm-3">
                     <label htmlFor="type" className="form-label">Descrição</label>
                     <input type="text" className="form-control" id="type" value={description} onChange={(e) => setDescription(e.target.value.toUpperCase())} />
+                </div>
+                <div className="col-sm-3">
+                    <label htmlFor="type" className="form-label">EAN</label>
+                    <input type="number" className="form-control" id="type" value={ean} onChange={(e) => setEAN(e.target.value)} />
                 </div>
                 <div className="col-sm-1">
                     <label htmlFor="quantity" className="form-label">Quantidade</label>
@@ -164,7 +228,7 @@ const Produtos: React.FC = () => {
                             </div>
                             <div className="col">
                                 <label htmlFor="editQuantity" className="form-label">Quantidade</label>
-                                <input type="number" className="form-control" id="editQuantity" value={quantity} onChange={(e) => setQuantity(parseFloat(e.target.value))} />
+                                <input type="text" className="form-control" id="editQuantity" value={quantity} onChange={(e) => setQuantity(parseFloat(e.target.value))} />
                             </div>
                             <div className="col">
                                 <label htmlFor="editUnitPrice" className="form-label">Valor Unitário</label>
@@ -175,22 +239,43 @@ const Produtos: React.FC = () => {
                                 <input type="text" className="form-control" id="editObs" value={obs} onChange={(e) => setObs(e.target.value.toUpperCase())} />
                             </div>
 
+                            <div className="col">
+                                <label htmlFor="editEAN" className="form-label">EAN</label>
+                                <input type="text" className="form-control" id="editEAN" value={ean} onChange={(e) => setEAN(e.target.value)} />
+
                         </div>
+                        </div>
+                        
                         <div className="modal-footer">
                             <button type="button" className="btn btn-primary" onClick={handleUpdateProduct} data-bs-dismiss="modal">Atualizar</button>
                         </div>
                     </div>
                 </div>
             </div>
-            <div>
-                <input type="text" className="form-control col my-3" placeholder="Filtrar por nome" onChange={handleFilterNameChange} />
+            <div className="d-flex gap-3 my-3">
+                <div>
+                    <input type="text" className="form-control" placeholder="Filtrar por nome" onChange={handleFilterNameChange} />
+                </div>
+                <div>
+                    <input type="text" className="form-control" placeholder="Filtrar por código" onChange={handleFilterCodeChange} />
+                </div>
+                <div className="form-check">
+                    <input className="form-check-input" type="checkbox" id="ocultarQtdZero" checked={ocultarZero} onChange={handleToggleOcultarZero} />
+                    <label className="form-check-label" htmlFor="ocultarQtdZero">Ocultar produtos com quantidade 0</label>
+                </div>
+                <div className="form-check">
+                    <input className="form-check-input" type="checkbox" id="exibirSomenteQtdZero" checked={exibeZero} onChange={handleToggleExibeZero} />
+                    <label className="form-check-label" htmlFor="exibirSomenteQtdZero">Exibir apenas produtos com quantidade 0</label>
+                </div>
+
             </div>
             <div style={{ height: '600px' }}>
                 <div className="h-100 overflow-y-auto border">
                     <table className="table table-striped table-bordered table-hover m-0">
                         <thead>
                             <tr>
-                                <th className="col">Code</th>
+                                <th className="col">Codigo</th>
+                                <th className="col">EAN</th>
                                 <th className="col">Descrição</th>
                                 <th className="col">Quantidade</th>
                                 <th className="col">Vl Unit</th>
@@ -200,27 +285,28 @@ const Produtos: React.FC = () => {
                             </tr>
                         </thead>
                         <tbody>
-                            {produtos.filter((product: Product) => product.description.toUpperCase().includes(filtroName.toUpperCase()))
-                                .map((product: Product) => (
-                                    <tr key={product.id} className="align-middle">
-                                        <td>{product.code}</td>
-                                        <td>{product.description}</td>
-                                        <td>{product.quantity}</td>
-                                        <td>R$ {product.unit_price}</td>
-                                        <td>{product.obs}</td>
-                                        <td>{product.registration_date}</td>
-                                        <td>
-                                            <button className="btn btn-primary me-3" data-bs-toggle="modal" data-bs-target="#staticBackdrop" onClick={() => handleEditProduct(product)}>Editar</button>
-                                            <button className="btn btn-danger me-3" onClick={() => handleDeleteProduct(product.id)}>Excluir</button>
-                                        </td>
-                                    </tr>
-                                ))}
+                            {produtosFiltrados.map((product: Product) => (
+                                <tr key={product.id} className="align-middle">
+                                    <td>{product.code}</td>
+                                    <td>{product.ean}</td>
+                                    <td>{product.description}</td>
+                                    <td>{product.quantity}</td>
+                                    <td>{formattedPrice(product.unit_price)}</td>
+                                    <td>{product.obs}</td>
+                                    <td>{product.registration_date}</td>
+                                    <td>
+                                        <button className="btn btn-primary me-3" data-bs-toggle="modal" data-bs-target="#staticBackdrop" onClick={() => handleEditProduct(product)}>Editar</button>
+                                        <button className="btn btn-danger me-3" onClick={() => handleDeleteProduct(product.id)}>Excluir</button>
+                                    </td>
+                                </tr>
+                            ))}
                         </tbody>
                     </table>
                 </div>
             </div>
             <div className="d-flex justify-content-center">
-                <button className="btn btn-success my-3" onClick={handleExportProducts}>Exportar</button>
+                <button className="btn btn-success my-3" onClick={handleExportProducts}>Baixar Planilha de Produtos</button>
+                <p>VALOR TOTAL: {valorTotal}</p>
             </div>
         </div>
     );
